Refactor: file-names to be more consistent
