+++ src/adapters/index.ts
@@ -2,7 +2,7 @@
 //  ADAPTERS
 //  --------
 
-export * from './Memory'
+export * from './memory'
 export * from './FS'
 export * from './localStorage'
 export * from './URL'
